# Faster env-map MIS calculation, use FloatType when loading env-maps with RGBELoader