--- a/example/skinnedMesh.js
+++ b/example/skinnedMesh.js
@@ -76,7 +76,7 @@ async function init() {
 	clock = new THREE.Clock();
 
 	// loading the
-	const envMapPromise = new RGBELoader()
+	const envMapPromise = new RGBELoader().setDataType( THREE.FloatType )
 		.loadAsync( 'https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/textures/equirectangular/royal_esplanade_1k.hdr' )
 		.then( texture => {