--- a/example/spotLights.js
+++ b/example/spotLights.js
@@ -112,7 +112,7 @@ async function init() {
 	samplesEl = document.getElementById( 'samples' );
 
 	// load the env map
-	const envMapPromise = new RGBELoader()
+	const envMapPromise = new RGBELoader().setDataType( THREE.FloatType )
 		.loadAsync( 'https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/textures/equirectangular/royal_esplanade_1k.hdr' )
 		.then( texture => {